+++ src/types/EtherspotTransactionKit.ts
@@ -1,7 +1,6 @@
-import { PaymasterApi as PaymasterApiModular } from '@etherspot/modular-sdk';
-import { PaymasterApi } from '@etherspot/prime-sdk';
-import { ExchangeOffer } from '@etherspot/prime-sdk/dist/sdk/data';
-import { TransactionStatuses } from '@etherspot/prime-sdk/dist/sdk/data/constants';
+import { ExchangeOffer } from '@etherspot/data-utils/dist/cjs/sdk/data/classes/exchange-offer';
+import { TransactionStatuses } from '@etherspot/data-utils/dist/cjs/sdk/data/constants';
+import { PaymasterApi } from '@etherspot/modular-sdk';
 import { Fragment, JsonFragment } from '@ethersproject/abi';
 import { Route } from '@lifi/types';
 import { BigNumber, BigNumberish, BytesLike } from 'ethers';
@@ -44,7 +43,7 @@ export interface IBatches {
   onEstimated?: (estimated: EstimatedBatch[]) => void;
   onSent?: (sent: SentBatch[]) => void;
   skip?: boolean;
-  paymaster?: PaymasterApi | PaymasterApiModular;
+  paymaster?: PaymasterApi;
 }
 
 export type IEstimatedBatches = IBatches & {
@@ -182,12 +181,6 @@ export interface UserOpTransaction {
   erc20Transfers?: EtherspotErc20TransfersEntity[];
   nftTransfers?: EtherspotNftTransfersEntity[];
 }
-
-export type AccountTemplate =
-  | 'etherspot'
-  | 'etherspotModular'
-  | 'zeroDev'
-  | 'simpleAccount';
 
 // eslint-disable-next-line @typescript-eslint/naming-convention
 export enum MODULE_TYPE {